upgrading precison to double

+++ headers/track/model/fast_bvh/IntersectionInfo.h
@@ -4,7 +4,7 @@
 struct Object;
 
 struct IntersectionInfo {
- float t; // Intersection distance along the ray
+ double t; // Intersection distance along the ray
  const Object* object; // Object that was hit
  Vector3 hit; // Location of the intersection
 };
