+++ src/track/model/fast_bvh/BVH.cpp
@@ -17,10 +17,12 @@ struct BVHTraversal {
 //! - In the case where we want to find out of there is _ANY_ intersection at all,
 //!   set occlusion == true, in which case we exit on the first hit, rather 
 //!   than find the closest.
+
+//swap IntersectionInfo for 2x this value, and search for both front and back
 bool BVH::getIntersection(const Ray& ray, IntersectionInfo* intersection, bool occlusion) const {
 	intersection->t = 999999999.f;
 	intersection->object = NULL;
- float bbhits[4];
+ double bbhits[4];
  int32_t closer, other;
 
  // Working set
